complet

+++ src/main/java/server/Server.java
@@ -3,8 +3,6 @@ package server;
 import javafx.util.Pair;
 import server.models.Course;
 import server.models.RegistrationForm;
-
-
 import java.io.*;
 import java.net.ServerSocket;
 import java.net.Socket;
@@ -13,8 +11,11 @@ import java.util.Arrays;
 import java.util.Objects;
 import java.util.Scanner;
 
-public class Server {
+/**
+ * Cette classe contient le code pour le serveur. Elle recoit les requetes et fait le traitement.
+ */
 
+public class Server {
     public final static String REGISTER_COMMAND = "INSCRIRE";
     public final static String LOAD_COMMAND = "CHARGER";
     private final ServerSocket server;
@@ -22,7 +23,13 @@ public class Server {
     private ObjectInputStream objectInputStream;
     private ObjectOutputStream objectOutputStream;
     private final ArrayList<EventHandler> handlers;
+    private ArrayList<Course> courseList = new ArrayList<Course>();
 
+    /**
+     * cette methode cree un serveur et gere les evenements.
+     * @param port numero du port sur lequel le serveur doit etre cree.
+     * @throws IOException
+     */
     public Server(int port) throws IOException {
         this.server = new ServerSocket(port, 1);
         this.handlers = new ArrayList<EventHandler>();
@@ -39,6 +46,9 @@ public class Server {
         }
     }
 
+    /**
+     * cette methode accepte les requetes des clients. Elle ecoute les commande envoyes et deconnecte le client apres l'execution
+     */
     public void run() {
         while (true) {
             try {
@@ -55,6 +65,12 @@ public class Server {
         }
     }
 
+    /**
+     * methode qui lit les commande des client et recupere les informations. Les commandes sont ensuite passées à
+     * alertHandlers qui poursuit la suite du traitement
+     * @throws IOException
+     * @throws ClassNotFoundException
+     */
     public void listen() throws IOException, ClassNotFoundException {
         String line;
         if ((line = this.objectInputStream.readObject().toString()) != null) {
@@ -65,6 +81,11 @@ public class Server {
         }
     }
 
+    /**
+     * methode qui separe une ligne de texte en commandes et argument a passer dans le programme.
+     * @param line information entrée par l'utilisateur une commande et un argument
+     * @return une commande et son argument sont retournées par la methode
+     */
     public Pair<String, String> processCommandLine(String line) {
         String[] parts = line.split(" ");
         String cmd = parts[0];
@@ -72,12 +93,21 @@ public class Server {
         return new Pair<>(cmd, args);
     }
 
+    /**
+     * methode qui deconnecte le client.
+     * @throws IOException
+     */
     public void disconnect() throws IOException {
         objectOutputStream.close();
         objectInputStream.close();
         client.close();
     }
 
+    /**
+     * methode qui appelle une fonction selon la commande entrée et l'argument donné en parametre
+     * @param cmd commande entree par l'utilisateur.
+     * @param arg session que l'utilisateur souhaite afficher.
+     */
     public void handleEvents(String cmd, String arg) {
         if (cmd.equals(REGISTER_COMMAND)) {
             handleRegistration();
@@ -100,10 +130,8 @@ public class Server {
     public void handleLoadCourses(String arg) {
         try {
 
-            ArrayList<Course> courseList = new ArrayList<Course>();
             arg = "";
             courseList.clear();
-
             arg = objectInputStream.readObject().toString();
             int valeur = Integer.parseInt(arg);
             if (!(valeur >= 1 && valeur <= 3))
@@ -115,6 +143,10 @@ public class Server {
             String jarPath = new File(getClass().getProtectionDomain().getCodeSource().getLocation().getPath()).getParent();
 
             File file = new File(jarPath + File.separator + "cours.txt");
+            if(!file.exists()){
+                throw new FileNotFoundException();
+            }
+
             Scanner ligne = new Scanner(new InputStreamReader(new FileInputStream(file)));
 
             while (ligne.hasNextLine()) {
@@ -129,12 +161,17 @@ public class Server {
 
                 }
             }
-
             objectOutputStream.writeObject(courseList);
             objectOutputStream.flush();
             courseList.clear();
         } catch (Exception e) {
-            e.printStackTrace();
+            try {
+
+                objectOutputStream.writeObject(courseList);
+            }catch (IOException f){
+                f.printStackTrace();
+
+            }
         }
     }
 
@@ -148,25 +185,32 @@ public class Server {
      */
     public void handleRegistration() {
         try {
-            Object formeRecu = objectInputStream.readObject();
-
-
-
+            RegistrationForm formeRecu = (RegistrationForm) objectInputStream.readObject();
             String jarPath = new File(getClass().getProtectionDomain().getCodeSource().getLocation().getPath()).getParent();
             File file = new File(jarPath + File.separator + "inscription.txt");
+            if(!file.exists()){
+                throw new FileNotFoundException();
+            }
+
             FileWriter writer = new FileWriter(file);
             PrintWriter pw = new PrintWriter(writer);
-            pw.println(formeRecu);
+            Course cours = formeRecu.getCourse();
+            String aEcrire = cours.getSession() + " \t" + cours.getCode() + " \t" + formeRecu.getMatricule() + " \t"+
+                    formeRecu.getPrenom()+ " \t"
+                    + formeRecu.getNom()+ " \t" + formeRecu.getEmail();
+            pw.println(aEcrire);
             pw.close();
             writer.close();
-
-
-
+            objectOutputStream.writeObject("Inscription reussite !");
 
 
 
         } catch (Exception e) {
-            throw new RuntimeException("Erreur lors de l'enregistrement de l'inscription", e);
+            try{
+            objectOutputStream.writeObject("Erreur lors de l'inscription !");
+
+        } catch (IOException l){
+            l.printStackTrace();}
         }
 
     }
